19 support tags for series

+++ src/main/java/com/springboot/cric/models/SeriesTeamsMap.java
@@ -14,6 +14,6 @@ public class SeriesTeamsMap {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Long seriesId;
+    private Integer seriesId;
     private Long teamId;
 }
