+++ src/main/java/com/springboot/cric/repositories/ManOfTheSeriesRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface ManOfTheSeriesRepository extends JpaRepository<ManOfTheSeries, Long> {
-    List<ManOfTheSeries> findAllBySeriesIdIn(List<Long> seriesIds);
-    List<ManOfTheSeries> findAllBySeriesIdAndPlayerIdIn(Long seriesId, List<Long> playerIds);
+    List<ManOfTheSeries> findAllBySeriesIdIn(List<Integer> seriesIds);
+    List<ManOfTheSeries> findAllBySeriesIdAndPlayerIdIn(Integer seriesId, List<Long> playerIds);
     List<ManOfTheSeries> findAllByPlayerId(Long playerId);
 }
