+++ src/main/java/com/springboot/cric/repositories/SeriesTeamsMapRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface SeriesTeamsMapRepository extends JpaRepository<SeriesTeamsMap, Long> {
-    List<SeriesTeamsMap> findAllBySeriesIdIn(List<Long> seriesIds);
-    List<SeriesTeamsMap> findAllBySeriesIdAndTeamIdIn(Long seriesId, List<Long> teamIds);
+    List<SeriesTeamsMap> findAllBySeriesIdIn(List<Integer> seriesIds);
+    List<SeriesTeamsMap> findAllBySeriesIdAndTeamIdIn(Integer seriesId, List<Long> teamIds);
 }
